+++ docs/changelog.md
@@ -10,6 +10,7 @@ pages](https://github.com/dalibo/ldap2pg/pulls?utf8=%E2%9C%93&q=is%3Apr%20is%3Am
 # Unreleased
 
 - Postgres 12 support validated.
+- Fix void attributes raising *Missing attribute error*.
 
 
 # ldap2pg 5.0